Create pubmaticUtils for utility functions
